Класс Policy (Политика прав пользователей), доработка маршрутизатора (Routing)

+++ app/Http/Controllers/WorkerController.php
@@ -6,12 +6,16 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\Worker\WorkerRequest;
 use App\Models\Worker;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Auth\Access\Response;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 
 class WorkerController extends Controller
 {
     /**
+     * @param WorkerRequest $request
+     *
      * @return View
      */
     public function index(WorkerRequest $request): View
@@ -66,19 +70,27 @@ class WorkerController extends Controller
 
     /**
      * @return View
+     * @throws AuthorizationException
      */
     public function create(): View
     {
+        // Проверка App\Policies\WorkerPolicy у пользователя
+        $this->authorize('create', Worker::class);
+
         return view('worker.create');
     }
 
     /**
      * @param WorkerRequest $request
      *
-     * @return RedirectResponse
+     * @return RedirectResponse|Response
+     * @throws AuthorizationException
      */
-    public function store(WorkerRequest $request): RedirectResponse
+    public function store(WorkerRequest $request): RedirectResponse|Response
     {
+        // Проверка App\Policies\WorkerPolicy у пользователя
+        $this->authorize('create', Worker::class);
+
         $worker = $request->validated();
         $worker['is_married'] = !empty($worker['is_married']);
         Worker::query()->create($worker);
@@ -90,9 +102,13 @@ class WorkerController extends Controller
      * @param Worker $worker
      *
      * @return View
+     * @throws AuthorizationException
      */
     public function edit(Worker $worker): View
     {
+        // Проверка App\Policies\WorkerPolicy у пользователя
+        $this->authorize('update', $worker);
+
         return view('worker.edit', compact('worker'));
     }
 
@@ -101,9 +117,13 @@ class WorkerController extends Controller
      * @param Worker $worker
      *
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(WorkerRequest $request, Worker $worker): RedirectResponse
     {
+        // Проверка App\Policies\WorkerPolicy у пользователя
+        $this->authorize('update', $worker);
+
         $data = $request->validated();
         $data['is_married'] = !empty($data['is_married']);
         $worker->update($data);
@@ -115,9 +135,13 @@ class WorkerController extends Controller
      * @param Worker $worker
      *
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function delete(Worker $worker): RedirectResponse
     {
+        // Проверка App\Policies\WorkerPolicy у пользователя
+        $this->authorize('delete', $worker);
+
         $worker->delete();
         return redirect()->route('workers.index')->with('success', 'Работник успешно удалён');
     }
